In the updated code, we have refactored the logout process on both frontend and backend for better handling different environments (development and production).
Frontend:
The GET request for logout now accounts for whether the application is in development or production stage. Previously, it would only redirect to '/logout' which wouldn't work as expected during development on localhost.

Backend:
The rejigged logic depends on the environment variable to direct the behavior of session termination. This environmental segregation was absent earlier which caused issues in the dev environment. We also introduced a new environment variable in the application.properties file for customizing behavior based on current environment.

.gitignore:
Added extra exclusion rules for the application secrets file and resources/static directory, giving us better control on what is committed to the repository.

In essence, this commit improves the handling of session termination and logout for different environments, providing a more robust and fitting solution for both development and production stages.